MUL-727: Refactoring.
Removed transaction_has_trait() API;
Moved BinaryData into binary_data.h/cpp;
Moved validate_address() to blockchain.h/cpp;
Added bitcoin.h, ethereum.h, golos.h that have blockchain-specific enums;
Split utility into error_utility and binary_data_utility;
A bit simplified error management.
