+++ multy_core/account.h
@@ -9,6 +9,8 @@
 
 #include "multy_core/api.h"
 
+#include "multy_core/blockchain.h"
+
 #include <stddef.h>
 #include <stdint.h>
 
@@ -21,26 +23,6 @@ struct Account;
 struct Error;
 struct ExtendedKey;
 struct Key;
-
-// See: https://github.com/satoshilabs/slips/blob/master/slip-0044.md
-enum Blockchain
-{
-    BLOCKCHAIN_BITCOIN = 0x00,
-    BLOCKCHAIN_ETHEREUM = 0x3c,
-    BLOCKCHAIN_GOLOS = 0x060105
-};
-
-enum BlockchainNetType
-{
-    BLOCKCHAIN_NET_TYPE_MAINNET = 0,
-    BLOCKCHAIN_NET_TYPE_TESTNET = 1,
-};
-
-struct BlockchainType
-{
-    enum Blockchain blockchain;
-    size_t net_type; // blockchain-specific net type, 0 for MAINNET.
-};
 
 enum AddressType
 {
@@ -128,16 +110,6 @@ MULTY_CORE_API struct Error* account_get_address_path(
 MULTY_CORE_API struct Error* account_get_blockchain_type(
         const struct Account* account,
         struct BlockchainType* out_blockchain_type);
-
-// TODO: move to blockchain header, with blockchain_get_string()
-/** Validate an address for given blockchain
- *  @param address - address
- *  @param blockchain_type - Blockchain to use address for.
- *  @return null ptr if address is valid, Error if it is not.
- */
-MULTY_CORE_API struct Error* validate_address(
-        struct BlockchainType blockchain,
-        const char* address);
 
 /** Frees HDAccount instance, can accept nullptr. **/
 MULTY_CORE_API void free_hd_account(struct HDAccount*);
