+++ multy_core/common.h
@@ -29,13 +29,6 @@ struct EntropySource
     size_t (*fill_entropy)(void* data, size_t size, void* dest);
 };
 
-/** Binary data, just a pointer and a size in bytes. */
-struct BinaryData
-{
-    const unsigned char* data;
-    size_t len;
-};
-
 struct Version
 {
     size_t major;
@@ -52,35 +45,6 @@ MULTY_CORE_API struct Error* get_version(struct Version* version);
  * @return Error on error, nullptr otherwise.
  */
 MULTY_CORE_API struct Error* make_version_string(const char** out_version_string);
-
-/** Frees BinaryData, can take null. **/
-MULTY_CORE_API void free_binarydata(struct BinaryData*);
-
-/** Create new BinaryData with data of given size, data is zeroed. **/
-MULTY_CORE_API struct Error* make_binary_data(
-        size_t size, struct BinaryData** new_binary_data);
-
-MULTY_CORE_API struct Error* make_binary_data_from_bytes(
-        const unsigned char* data, size_t size,
-        struct BinaryData** new_binary_data);
-
-MULTY_CORE_API struct Error* make_binary_data_from_hex(
-        const char* hex_str, struct BinaryData** new_binary_data);
-
-/** Copies BinaryData. **/
-MULTY_CORE_API struct Error* make_binary_data(
-        size_t size, struct BinaryData** new_binary_data);
-
-MULTY_CORE_API struct Error* make_binary_data_from_bytes(
-        const unsigned char* data, size_t size,
-        struct BinaryData** new_binary_data);
-
-MULTY_CORE_API struct Error* make_binary_data_from_hex(
-        const char* hex_str, struct BinaryData** new_binary_data);
-
-/** Copies BinaryData. **/
-MULTY_CORE_API struct Error* make_binary_data_clone(
-        const struct BinaryData* source, struct BinaryData** new_binary_data);
 
 /** Frees a string, can take null. **/
 MULTY_CORE_API void free_string(const char* str);
