+++ multy_core/golos.h
@@ -1,0 +1,24 @@
+/* Copyright 2018 by Multy.io
+ * Licensed under Multy.io license.
+ *
+ * See LICENSE for details
+ */
+
+#ifndef MULTY_CORE_GOLOS_H
+#define MULTY_CORE_GOLOS_H
+
+#ifdef __cplusplus
+extern "C" {
+#endif
+
+enum GolosNetType
+{
+    GOLOS_MAINNET = 0,
+    GOLOS_TESTNET = 1,
+};
+
+#ifdef __cplusplus
+} // extern "C"
+#endif
+
+#endif /* MULTY_CORE_GOLOS_H */
