+++ multy_core/src/api/account.cpp
@@ -215,49 +215,6 @@ Error* account_get_blockchain_type(
     return nullptr;
 }
 
-Error* validate_address(BlockchainType blockchain_type, const char* address)
-{
-    ARG_CHECK(address);
-    ARG_CHECK(blockchain_type.blockchain == BLOCKCHAIN_BITCOIN
-            || blockchain_type.blockchain == BLOCKCHAIN_ETHEREUM);
-
-    try
-    {
-        switch (blockchain_type.blockchain)
-        {
-            case BLOCKCHAIN_BITCOIN:
-            {
-                BitcoinAddressType address_type;
-                BitcoinNetType net_type;
-                bitcoin_parse_address(address, &net_type, &address_type);
-                if (address_type != BITCOIN_ADDRESS_P2PKH)
-                {
-                    THROW_EXCEPTION("BTC: Only P2PKH addresses are supported for now.");
-                }
-                if (net_type != blockchain_type.net_type)
-                {
-                    THROW_EXCEPTION("Incompatitable net_type.")
-                            << " Requested: " << blockchain_type.net_type
-                            << ", address net type:" << net_type;
-                }
-                break;
-            }
-            case BLOCKCHAIN_ETHEREUM:
-            {
-                THROW_EXCEPTION("ETH addresses are not supported yet.");
-            }
-            default:
-            {
-                return MAKE_ERROR(
-                        ERROR_GENERAL_ERROR, "Blockchain not supported yet");
-            }
-        }
-    }
-    CATCH_EXCEPTION_RETURN_ERROR();
-
-    return nullptr;
-}
-
 void free_hd_account(HDAccount* account)
 {
     CHECK_OBJECT_BEFORE_FREE(account);
