+++ multy_core/src/api/error.cpp
@@ -8,43 +8,47 @@
 
 #include "multy_core/common.h"
 
-#include "wally_core.h"
+#include "multy_core/src/error_utility.h"
 
 #include <stdlib.h>
 
 namespace
 {
+// Can't be const because make_error returns non-const Error.
+Error OutOfMemoryError;
 
-ErrorCode convert_error_code(int code)
+// Just in case, resetting OutOfMemory before passing it to the user.
+Error* get_out_of_memory_error()
 {
-    switch (code)
+    OutOfMemoryError = Error
     {
-        case WALLY_EINVAL:
-            return ERROR_INTERNAL;
-        case WALLY_ENOMEM:
-            return ERROR_OUT_OF_MEMORY;
-        case WALLY_OK:
-        case WALLY_ERROR:
-        default:
-            return ERROR_GENERAL_ERROR;
-    }
+        ERROR_OUT_OF_MEMORY,
+        "Failed to allocate Error object.",
+        false,
+        MULTY_CODE_LOCATION
+    };
+    return &OutOfMemoryError;
 }
 
 } // namespace
 
 Error* make_error(ErrorCode code, const char* message, CodeLocation location)
 {
-    return new Error{code, message, false, location};
-}
-
-Error* internal_make_error(int code, const char* message, CodeLocation location)
-{
-    return make_error(convert_error_code(code), message, location);
+    message = message ? message : "";
+    try
+    {
+        return new Error{code, message, false, location};
+    }
+    catch(...)
+    {
+        return get_out_of_memory_error();
+    }
+    return nullptr;
 }
 
 void free_error(Error* error)
 {
-    if (!error)
+    if (!error || error == &OutOfMemoryError)
     {
         return;
     }
@@ -52,6 +56,5 @@ void free_error(Error* error)
     {
         free_string(error->message);
     }
-    // no need to delete error message, since it is always a static string.
     delete error;
 }
