+++ multy_core/src/utility.cpp
@@ -20,63 +20,13 @@
 namespace
 {
 using namespace multy_core::internal;
-
-class ErrorWrapperException : public Exception
-{
-public:
-    ErrorWrapperException(ErrorPtr error)
-        : Exception("", error ? error->location : CodeLocation{nullptr, 0}),
-          m_error(std::move(error))
-    {
-        assert(m_error.get());
-    }
-
-    ErrorWrapperException(Error* error)
-        : Exception("", error ? error->location : CodeLocation{nullptr, 0}),
-          m_error(std::move(error))
-    {
-        assert(m_error.get());
-    }
-
-    const char* what() const noexcept override
-    {
-        return m_error->message;
-    }
-
-    Error* make_error() const override
-    {
-        return m_error.release();
-    }
-
-private:
-    mutable ErrorPtr m_error;
-};
-
-Error* make_error_from_string(const char* message)
-{
-    // Copy a message from exception to a new buffer.
-    // Buffer memory has to be allocated by libwally, since we want
-    // to use free_string() consistently on all strings,
-    // even if those were allocated internaly.
-    const char* message_copy = copy_string(message);
-    if (message && !message_copy)
-    {
-        // We have to guess here, most likely out of memory.
-        return MAKE_ERROR(ERROR_OUT_OF_MEMORY, "Out of memory");
-    }
-
-    Error* error = MAKE_ERROR(ERROR_GENERAL_ERROR, message_copy);
-    error->owns_message = true;
-    return error;
-}
-
 } // namespace
 
 namespace multy_core
 {
-
 namespace internal
 {
+
 char* copy_string(const std::string& str)
 {
     return copy_string(str.c_str());
@@ -102,7 +52,7 @@ char* copy_string(const char* str)
     if (!new_message)
     {
         THROW_EXCEPTION("Failed to allocate memory.")
-                << " requested: " << len + 1;
+                << " Requested: " << len + 1;
     }
 
     memcpy(new_message, str, len);
@@ -110,113 +60,5 @@ char* copy_string(const char* str)
     return new_message;
 }
 
-BinaryData power_slice(const BinaryData& data, int32_t offset, int32_t size)
-{
-    if (offset < 0)
-    {
-        offset += data.len;
-    }
-    if (size < 0)
-    {
-        size += data.len;
-    }
-    if (offset < 0 || size < 0)
-    {
-        THROW_EXCEPTION("Can't power_slice BinaryData: either offset or size "
-                "is too small.")
-                << " data length: " << data.len
-                << " offset: " << offset
-                << " size: " << size;
-    }
-
-    return slice(data, offset, size);
-}
-
-BinaryData slice(const BinaryData& data, size_t offset, size_t size)
-{
-    if (!data.data)
-    {
-        THROW_EXCEPTION("Can't slice BinaryData: data is nullptr.");
-    }
-    if (offset > data.len)
-    {
-        THROW_EXCEPTION("Can't slice BinaryData: offset > data length.")
-                << " data length: " << data.len
-                << " offset: " << offset;
-    }
-    if (offset + size > data.len)
-    {
-        THROW_EXCEPTION("Can't slice BinaryData: offset + size > data length.")
-                << " data length: " << data.len
-                << " offset: " << offset
-                << " size: " << size;
-    }
-
-    return BinaryData{data.data + offset, size};
-}
-
-BinaryData as_binary_data(const char* str)
-{
-    const size_t len = str ? strlen(str) : 0;
-    return BinaryData{reinterpret_cast<const unsigned char*>(str), len};
-}
-
-void throw_if_error(Error* error)
-{
-    if (error)
-    {
-        throw ErrorWrapperException(error);
-    }
-}
-
-#ifdef MULTY_ENABLE_SIMULATE_ERROR
-int simulate_error(int err_code, const char* /*statement*/, const CodeLocation& /*location*/)
-{
-    // Here you can inject an error for some statement and\or code location,
-    // simulating specific failure.
-    if (false)
-    {
-        return INT32_MIN;
-    }
-    return err_code;
-}
-#endif
-
-void throw_if_wally_error(int err_code, const char* message, const CodeLocation& location)
-{
-    if (err_code != 0)
-    {
-        throw_if_error(internal_make_error(err_code, message, location));
-    }
-}
-Error* exception_to_error(const CodeLocation& location)
-{
-    try
-    {
-        throw;
-    }
-    catch (const multy_core::internal::Exception& exception)
-    {
-        return exception.make_error();
-    }
-    catch (const std::exception& exception)
-    {
-        Error* result = make_error_from_string(exception.what());
-        result->location = location;
-        return result;
-    }
-    catch (...)
-    {
-        return make_error(ERROR_GENERAL_ERROR, "Unknown exception", location);
-    }
-}
-
-bool operator==(const BinaryData& left, const BinaryData& right)
-{
-    return left.len == right.len
-            && memcmp(left.data, right.data, left.len) == 0;
-}
-
 } // namespace internal
-
 } // namespace multy_core
